pyproject.toml
[project]
name = "kielas-rm-detector-train"
version = "0.1.0"
description = "适合 rm 视觉体质的关键点检测模型"
readme = "README.md"
authors = [
    { name = "Kielas", email = "c1470759@outlook.com" }
]

# 将 "==3.10" 修改为 ">=3.10"，以提供更好的环境兼容性。
# 如果你必须严格锁定 3.10 版本，可以改回 "==3.10"。
requires-python = ">=3.10"

dependencies = [
    "cvbridge3>=1.1",
    "matplotlib>=3.10.*",
    "numpy>=1.24.0,<2.0.0",
    "opencv-python<4.11",  # 限制版本以兼容 NumPy 1.x
    "rich>=14.3.3",
    "PyYAML",
    "requests>=2.33.1",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "ipywidgets>=8.0.0",
    # "tensorboard>=2.17.0", # [可选] 推荐添加，用于后续在训练中记录 Loss 和精度曲线
]

[project.optional-dependencies]
deploy = [
    "onnx>=1.16.0",
    "onnxscript>=0.1.0",
    "onnxsim>=0.4.33",
    "torchsummary>=1.5.1",
    "onnxruntime-gpu>=1.18.0", # 替换为 GPU 版本
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
extra-index-url = ["https://download.pytorch.org/whl/cu124"]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true